+++ plugins/taskbar/taskbar.c
@@ -78,7 +78,7 @@ taskbar_destructor(plugin_instance *p)
 {
     taskbar_priv *tb = (taskbar_priv *) p;
 
-    gdk_window_remove_filter(NULL, (GdkFilterFunc)tb_event_filter, tb);
+    /* Per-window filters are removed in del_task via task_remove_every below. */
     g_signal_handlers_disconnect_by_func(G_OBJECT (fbev),
             tb_net_current_desktop, tb);
     g_signal_handlers_disconnect_by_func(G_OBJECT (fbev),